backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.0
python-multipart==0.0.6
pillow==10.0.1
numpy==1.26.0
pyttsx3==2.90
pytesseract==0.3.10
python-dotenv==1.0.0